fix(project): Avoid using commons names , clougence-commons -> clougence-utils

+++ clougence-schema/src/main/java/com/clougence/schema/metadata/domain/rdb/adb/mysql/AdbMySqlColumn.java
@@ -47,7 +47,7 @@ public class AdbMySqlColumn implements ColumnDef {
     private String                      defaultCollationName;
     private String                      defaultCharacterSetName;
     private Long                        charactersMaxLength;
-    private Integer                     bytesMaxLength;
+    private Long                        bytesMaxLength;
     //
     private AdbMySqlOnCurrentUpdateType onCurrentUpdateType;
